refactor: 컨트롤러 패키지 위치를 com 밑으로 변경
