+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,8 +1,9 @@
-package org.apache.coyote.http11.controller;
+package com.techcourse.controller;
 
 import java.io.IOException;
 import java.util.Map;
 
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
