+++ tomcat/src/main/java/com/techcourse/controller/StaticResourceController.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.http11.controller;
+package com.techcourse.controller;
 
 import java.io.IOException;
 
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
